registering the service before registering the contribution
